REFACTOR : vcReqDto reqDays, gvDto remainDays 를 long에서 double 타입으로 변경 (반차 사용 때문에)

+++ src/main/java/com/meta/ale/domain/GrantedVcDto.java
@@ -20,7 +20,7 @@ public class GrantedVcDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private Date expiredDate;
     private Long vcDays;
-    private Long remainDays;
+    private Double remainDays;
     private VcTypeDto vcTypeDto;
     private EmpDto empDto;
 }
